+++ crucible-llvm/src/Lang/Crucible/LLVM/Intrinsics/Libc.hs
@@ -99,6 +99,7 @@ libc_overrides =
   , SomeLLVMOverride llvmStrcpyOverride
   , SomeLLVMOverride llvmStrdupOverride
   , SomeLLVMOverride llvmStrndupOverride
+  , SomeLLVMOverride llvmMemcmpOverride
   , SomeLLVMOverride llvmPrintfOverride
   , SomeLLVMOverride llvmPrintfChkOverride
   , SomeLLVMOverride llvmPutsOverride
@@ -428,6 +429,14 @@ llvmStrndupOverride =
   [llvmOvr| i8* @strndup( i8*, size_t ) |]
   (\memOps args -> Ctx.uncurryAssignment (callStrndup memOps) args)
 
+llvmMemcmpOverride
+  :: ( IsSymInterface sym, HasLLVMAnn sym, HasPtrWidth wptr
+     , ?memOpts :: MemOptions )
+  => LLVMOverride p sym ext (EmptyCtx ::> LLVMPointerType wptr ::> LLVMPointerType wptr ::> BVType wptr) (BVType 32)
+llvmMemcmpOverride =
+  [llvmOvr| i32 @memcmp( i8*, i8*, size_t ) |]
+  (\memOps args -> Ctx.uncurryAssignment (callMemcmp memOps) args)
+
 ------------------------------------------------------------------------
 -- ** Implementations
 
@@ -709,6 +718,19 @@ callStrndup mvar (regValue -> src) (regValue -> bound) =
         Just b ->
           let bound' = Just (fromIntegral b) in
           CStr.dupConcretelyNullTerminatedString bak mem src bound' loc' noAlignment
+
+callMemcmp
+  :: ( IsSymInterface sym, HasPtrWidth wptr, HasLLVMAnn sym
+     , ?memOpts :: MemOptions )
+  => GlobalVar Mem
+  -> RegEntry sym (LLVMPointerType wptr)
+  -> RegEntry sym (LLVMPointerType wptr)
+  -> RegEntry sym (BVType wptr)
+  -> OverrideSim p sym ext r args ret (RegValue sym (BVType 32))
+callMemcmp mvar (regValue -> ptr1) (regValue -> ptr2) (regValue -> len) =
+  ovrWithBackend $ \bak -> do
+    mem <- readGlobal mvar
+    liftIO $ CStr.memcmp bak mem ptr1 ptr2 len
 
 callAssert
   :: ( IsSymInterface sym, HasPtrWidth wptr, HasLLVMAnn sym