+++ crucible-llvm/CHANGELOG.md
@@ -42,8 +42,9 @@
   * `Lang.Crucible.LLVM.MemModel.ppLLVMIntrinsicTypes`
   * `Lang.Crucible.LLVM.MemModel.ppLLVMMemIntrinsicType`
   * `Lang.Crucible.LLVM.MemModel.Pointer.ppLLVMPointerIntrinsicType`
-* Overrides for `strnlen`, `strcpy`, `strdup`, and `strndup` supported by new
+* Overrides for `strnlen`, `strcpy`, `strdup`, and `strndup`, supported by new
   APIs in `Lang.Crucible.LLVM.MemModel.Strings`.
+* Override for `memcmp`.
 
 # 0.8.0 -- 2025-11-09